requirement.txt
psycopg2~=2.9.3
paho-mqtt~=1.6.1
pandas~=1.4.3
setuptools~=63.2.0